src/main/java/io/aiven/kafka/connect/gcs/gcs/GoogleCredentialsBuilder.java
/*
 * Aiven Kafka GCS Connector
 * Copyright (c) 2019 Aiven Ltd
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package io.aiven.kafka.connect.gcs.gcs;

import com.google.auth.oauth2.GoogleCredentials;

import java.io.FileInputStream;
import java.io.IOException;
import java.io.InputStream;

public final class GoogleCredentialsBuilder {
    public static GoogleCredentials build(final String credentialsPath) throws IOException {
        if (credentialsPath == null) {
            return GoogleCredentials.getApplicationDefault();
        } else {
            try (final InputStream stream = new FileInputStream(credentialsPath)) {
                return GoogleCredentials.fromStream(stream);
            } catch (final IOException e) {
                throw new IOException("Failed to read GCS credentials from " + credentialsPath, e);
            }
        }
    }
}
